Handle comparison errors in "sort" filter
It is not guaranteed that all objects in a sortable list are of the same
type. We currently assume that all objects are comparable via
Comparable<Object>. This unfortunately is not the case, as most
implementations only allow comparing to other instances of the same
type (e.g., String vs. String, not String vs. CharSequence), throwing a
ClassCastException upon mismatch.

Use a custom comparator that handles this case, first trying to compare
the other way around (assuming that the custom RenderTransformer object
implementation can indeed compare against Strings), falling back to
comparing the outputs of toString() for the two objects.

An on-demand toString() cache is used to handle duplicate toString()
conversions.